feat: Use sqlalchemy's event system to apply redshift behavior.
Using patches/manually wrapping an engine's `execute` method is
the existing behavior and is comparatively fragile. Instead we now
use sqlalchemy's event system (which is designed for this sort of
thing!) to hook into the query execution to add our behavior.

This fixes long-standing inconsistencies between default sqlalchemy
behavior, and hopefully better hides the changes to calling code.