feat: Perform container cleanup in a multiprocess safe way.
Implements a more sophisticated conatiner cleanup mechanism which is
supposed to be multiprocess-safe.

Today, if you run tests in multiprocess mode with python-xdist, you
run the risk that the first test process to create the container
"wins" by creating the container. Then later when tests are about
to complete, that specific process might run out of tests first and
kill the container will other processes are still running (causing
test failures)

Instead, this implementation writes the container ids (produced by
whichever process “wins”) to a process-safe lock file and then stop
the container in the root process after tests complete