fix: Move port selection into the filelock.
This is primarily useful when `port` is set to `None`, leading to random
port selction for the scope of fixture being used.

The expectation is that for tests not using this feature, the port
selection is a simple lookup and therefore the lock contention will only
apply 1 or (-n N) N times, such that the differnce to everyone else is
negligible.