fix: Yield config rather than the container.
Yielding the container was just something we did originally, likely
because it was "better" than yielding nothing. However we dont actually
make use of the container itself anywhere downstream...because you cant!
There's no way to guarantee that your test is the one that produces the
container, which means most of the time, you'll end up getting `None`!

So instead, we should yield the config. This is **actually** useful,
because the config value can be variable (think database name, or port)
depending on your settings.

The practical application here is that a 3rd party resource is more
easily able to "just make" calls to `get_container` and get something
useful back (ft https://github.com/DanCardin/responsaas).