feat: Add `cleanup_database` option to postgres fixture.
Note, this could **potentially** reveal preexisting issues in code
under test, that could be perceived as a "breaking" change.

By deleting the database under test at the end of the test's execution,
any database connections left connected to the database might cause the
`DELETE DATABASE` command to fail.

PMR will **try** to use the `WITH FORCE` option on database versions >=
13.0, but that option does not exist on prior versions of postgres.

In any case, if this happens, it **is** ultimately revealing a "bug"
in the code it is testing. Additionally, you can simply turn off
database cleanup in one of various ways.